Use Long for item ID

+++ app/src/main/java/dev/tanoc/android/stockin/EditItemActivity.kt
@@ -20,7 +20,7 @@ class EditItemActivity : ComponentActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
 
-        intent.getIntExtra("id", 0).let {
+        intent.getLongExtra("id", 0).let {
             model.updateId(it)
         }
         intent.getStringExtra("title")?.let {
