+++ app/src/main/java/dev/tanoc/android/stockin/model/Item.kt
@@ -1,3 +1,3 @@
 package dev.tanoc.android.stockin.model;
 
-data class Item(val id: Int, var title: String, var url: String)
+data class Item(val id: Long, var title: String, var url: String)
